+++ lib/features/copy-paste/ChoreoModdleCopy.js
@@ -1,192 +1,98 @@
+import ModdleCopy from 'bpmn-js/lib/features/copy-paste/ModdleCopy';
+import inherits from 'inherits';
 import {
-  forEach,
-  isArray,
-  isDefined,
-  isObject,
-  reduce,
-  has
-} from 'min-dash';
-
-
+  is
+} from 'bpmn-js/lib/util/ModelUtil';
 
 /**
- * Utility for copying model properties from source element to target element.
- *
+ * This class overwrites two methods the ModdleCopy in bpmn-js. This mainly fixes copying messageFlows and participantRefs.
+ * We also change the behavior from bpmn-js insofar that choreoActivities are actually copied on copy not on paste,
+ * preventing modification after the copy.
+ * @param {Injector} injector
+ * @param {Moddle} moddle
  * @param {EventBus} eventBus
  * @param {BpmnFactory} bpmnFactory
- * @param {BpmnModdle} moddle
+ * @constructor
  */
-export default function ChoreoModdleCopy(eventBus, bpmnFactory, moddle) {
-  this._bpmnFactory = bpmnFactory;
-  this._eventBus = eventBus;
+export default function ChoreoModdleCopy(injector, moddle, eventBus, bpmnFactory) {
   this._moddle = moddle;
+  injector.invoke(ModdleCopy, this);
+  const LOWER_PRIO = 650; // Must be lower than LOW_PRIORITY in BpmnCopyPaste.js
 
+  // In bpmn-js these event listeners are attached in BpmnCopyPaste which shares the module with ModdleCopy,
+  // however, they are not really decoupled anyways. To avoid having to overwrite that class as well, we simply attache
+  // it here.
+  eventBus.on('copyPaste.copyElement', LOWER_PRIO, (context) => {
+    const descriptor = context.descriptor;
+    const element = context.element;
+    const businessObject = descriptor.oldBusinessObject;
+
+    if (descriptor.isExpanded) {
+      descriptor.collapsed = false;
+    }
+    if (descriptor.isExpanded === false) {
+      descriptor.collapsed = true;
+    }
+    descriptor.hidden = element.hidden;
+
+    // We need to trick a bit with the messageFlowRef. We store the name of the messages—new messages will be created
+    // later by the CreateChoreoTaskBehavior—because its the only important info about messages which are not copied themselves.
+    // We also store direction and visibility. Unfortunately, context hints does not work or we would have to change much more code.
+    var copyBusinessObject = bpmnFactory.create(businessObject.$type);
+    if (businessObject.messageFlowRef) {
+      copyBusinessObject.messageHints = businessObject.messageFlowRef.map(
+        mf => {
+          return {
+            sourceRef: mf.sourceRef,
+            name: mf.messageRef.name,
+            isVisible: element.bandShapes.find(s => s.businessObject === mf.sourceRef).diBand.isMessageVisible
+          };
+        });
+    }
+
+    // To avoid that objects that were copied get manipulated and then the manipulated version gets pasted
+    // we already do a copy here.
+    // This is a known issue with bpmn-js https://github.com/bpmn-io/bpmn-js/issues/798
+    if (is(businessObject, 'bpmn:ChoreographyActivity')) {
+      descriptor.oldBusinessObject = this.copyElement(
+        businessObject,
+        copyBusinessObject
+      );
+    }
+  });
 
 }
+inherits(ChoreoModdleCopy, ModdleCopy);
 
 ChoreoModdleCopy.$inject = [
+  'injector',
+  'moddle',
   'eventBus',
-  'bpmnFactory',
-  'moddle'
+  'bpmnFactory'
 ];
 
-/**
- * Copy model properties of source element to target element.
- *
- * @param {ModdleElement} sourceElement
- * @param {ModdleElement} targetElement
- * @param {Array<string>} [propertyNames]
- *
- * @param {ModdleElement}
- */
-ChoreoModdleCopy.prototype.copyElement = function(sourceElement, targetElement, propertyNames) {
-  var self = this;
-
-  if (propertyNames && !isArray(propertyNames)) {
-    propertyNames = [ propertyNames ];
-  }
-
-  propertyNames = propertyNames || getPropertyNames(sourceElement.$descriptor);
-
-  var canCopyProperties = this._eventBus.fire('moddleCopy.canCopyProperties', {
-    propertyNames: propertyNames,
-    sourceElement: sourceElement,
-    targetElement: targetElement
-  });
-
-  if (canCopyProperties === false) {
-    return targetElement;
-  }
-
-  if (isArray(canCopyProperties)) {
-    propertyNames = canCopyProperties;
-  }
-
-  // copy properties
-  forEach(propertyNames, function(propertyName) {
-    var sourceProperty;
-
-    if (has(sourceElement, propertyName)) {
-      sourceProperty = sourceElement.get(propertyName);
-    }
-
-    var copiedProperty = self.copyProperty(sourceProperty, targetElement, propertyName);
-
-    var canSetProperty = self._eventBus.fire('moddleCopy.canSetCopiedProperty', {
-      parent: parent,
-      property: copiedProperty,
-      propertyName: propertyName
-    });
-
-    if (canSetProperty === false) {
-      return;
-    }
-
-    if (isDefined(copiedProperty)) {
-      targetElement.set(propertyName, copiedProperty);
-    }
-  });
-
-  return targetElement;
-};
-
-/**
- * Copy model property.
- *
- * @param {*} property
- * @param {ModdleElement} parent
- * @param {string} propertyName
- *
- * @returns {*}
- */
 ChoreoModdleCopy.prototype.copyProperty = function(property, parent, propertyName) {
-  var self = this;
+  let copiedProperty = ModdleCopy.prototype.copyProperty.call(this, property, parent, propertyName);
+  const propertyDescriptor = this._moddle.getPropertyDescriptor(parent, propertyName);
 
-  // allow others to copy property
-  var copiedProperty = this._eventBus.fire('moddleCopy.canCopyProperty', {
-    parent: parent,
-    property: property,
-    propertyName: propertyName
-  });
-
-  // return if copying is NOT allowed
-  if (copiedProperty === false) {
-    return;
-  }
-
-  if (copiedProperty) {
-    if (isObject(copiedProperty) && copiedProperty.$type && !copiedProperty.$parent) {
-      copiedProperty.$parent = parent;
+  // References are not copied by the bpmn-js ModdleCopy. However, we need to copy these references to reconstruct
+  // the Choreo Activity.
+  // Unlike most other references they do not require the creation of new objects during copy-paste.
+  if (!copiedProperty && propertyDescriptor.isReference) {
+    if (propertyDescriptor.name === 'participantRef') {
+      return [].concat(property);
     }
-
-    return copiedProperty;
-  }
-
-  var propertyDescriptor = this._moddle.getPropertyDescriptor(parent, propertyName);
-
-  // Chor-js specific
-  if (propertyDescriptor.name === 'participantRef') {
-    return [].concat(property);
-  }
-
-  if (propertyDescriptor.name === 'initiatingParticipantRef') {
-    return property;
-  }
-  // end chor-js specific
-
-  // do NOT copy Ids and references
-  if (propertyDescriptor.isId || propertyDescriptor.isReference) {
-    return;
-  }
-
-  // copy arrays
-  if (isArray(property)) {
-    return reduce(property, function(childProperties, childProperty) {
-
-      // recursion
-      copiedProperty = self.copyProperty(childProperty, parent, propertyName);
-
-      // copying might NOT be allowed
-      if (copiedProperty) {
-        copiedProperty.$parent = parent;
-
-        return childProperties.concat(copiedProperty);
-      }
-
-      return childProperties;
-    }, []);
-  }
-
-  // copy model elements
-  if (isObject(property) && property.$type) {
-    if (this._moddle.getElementDescriptor(property).isGeneric) {
-      return;
+    if (propertyDescriptor.name === 'initiatingParticipantRef') {
+      return property;
     }
-
-    copiedProperty = self._bpmnFactory.create(property.$type);
-
-    copiedProperty.$parent = parent;
-
-    // recursion
-    copiedProperty = self.copyElement(property, copiedProperty);
-
-    return copiedProperty;
   }
-
-  // copy primitive properties
-  return property;
+  return copiedProperty;
 };
 
-// helpers //////////
-
-export function getPropertyNames(descriptor, keepDefaultProperties) {
-  return reduce(descriptor.properties, function(properties, property) {
-
-    if (keepDefaultProperties && property.default) {
-      return properties;
-    }
-
-    return properties.concat(property.name);
-  }, []);
-}
+ChoreoModdleCopy.prototype.copyElement = function(sourceElement, targetElement, propertyNames) {
+  if (sourceElement.messageHints) {
+    targetElement.messageHints = sourceElement.messageHints;
+  }
+  return ModdleCopy.prototype.copyElement.call(this, sourceElement, targetElement, propertyNames);
+};
