+++ lib/features/copy-paste/index.js
@@ -1,16 +1,8 @@
 import CopyPasteModule from 'diagram-js/lib/features/copy-paste';
 
-import ChoreoCopyPaste from './ChoreoCopyPaste';
-import ChoreoModdleCopy from './ChoreoModdleCopy';
 import BpmnCopyPaste from 'bpmn-js/lib/features/copy-paste/BpmnCopyPaste';
-import NewChoreoModdleCopy from './NewChoreoModdleCopy';
+import ChoreoModdleCopy from './ChoreoModdleCopy';
 
-/**
- * We completely overwrite the standard bpmnCopyPaste and register ChoreoCopyPaste as its replacement. We needt to do this,
- * because bpmnCopyPaste registers a function in its constructor
- * EventBus.on('copyPaste.copyElement') that we do not want to execute an cannot otherwise 'nicely' overwrite and prevent from
- * bpmnCopyPaste from executing on paste.
- */
 export default {
   __depends__: [
     CopyPasteModule
@@ -19,7 +11,8 @@ export default {
     'moddleCopy',
     'bpmnCopyPaste'
   ],
-  moddleCopy: ['type', NewChoreoModdleCopy],
- // bpmnCopyPaste: ['type', ChoreoCopyPaste],
-  bpmnCopyPaste: ['type', BpmnCopyPaste],
+  moddleCopy: ['type', ChoreoModdleCopy],
+  bpmnCopyPaste: ['type', BpmnCopyPaste]
+  // We can use the original BpmnCopyPaste, as we add a low prio event listener in ChoreoModdleCopy that takes care
+  // of the chore specific things we need to handle.
 };
